requirements.txt
absl-py==1.4.0
beautifulsoup4==4.12.2
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
contourpy==1.0.7
cycler==0.11.0
filelock==3.11.0
fonttools==4.39.3
gdown==4.7.1
google-auth==2.17.3
google-auth-oauthlib==1.0.0
grpcio==1.53.0
idna==3.4
imageio==2.27.0
imgaug==0.4.0
importlib-metadata==6.4.1
importlib-resources==5.12.0
kiwisolver==1.4.4
lazy_loader==0.2
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
networkx==3.1
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
opencv-python==4.7.0.72
packaging==23.1
pandas==2.0.0
Pillow==9.5.0
protobuf==4.22.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.20.0
scipy==1.9.1
seaborn==0.12.2
shapely==2.0.1
six==1.16.0
soupsieve==2.4.1
tensorboard==2.12.2
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tifffile==2023.4.12
torch==1.13.1
torchcam==0.3.2
torchvision==0.14.1
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
Werkzeug==2.2.3
zipp==3.15.0
